+++ src/detail_src/addReview2.js
@@ -1,9 +1,9 @@
-import { beginToFirebase } from "./beginToFirebase.js";
+import { beginToFirebase } from "./beginToFirebase2.js";
 import {
   collection,
   addDoc,
 } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
-import { getCurrentDate } from "./getCurrentDate.js";
+import { getCurrentDate } from "./getCurrentDate2.js";
 
 // TODO: detailMain.js 파일에 함수 임포트 및 호출해야 함 -> O
 // TODO: 이때, detailInfo() 함수로부터 movieId(movie.id) 받아와야 함 -> O
