+++ src/detail_src/detailMain.js
@@ -1,11 +1,10 @@
 import { searchMovies } from "../all_src/searchMovies.js";
 import { toTheTop } from "../all_src/toTheTop.js";
-import { review } from "./review.js";
-// import { reviewCopy } from "./reviewCopy.js";
+// import { review } from "./review.js";
+import { review } from "./reviewCopy.js";
 import { detailInfo } from "./detailInfo.js";
 
 searchMovies();
 toTheTop();
 detailInfo();
 review();
-// reviewCopy();
