+++ src/detail_src/getReview2.js
@@ -1,5 +1,8 @@
 import { beginToFirebase } from "./beginToFirebase2.js";
-import { collection, getDocs } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
+import {
+  collection,
+  getDocs,
+} from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
 
 // TODO: detailMain.js 파일에 함수 임포트 및 호출해야 함 -> O
 // TODO: 이때, detailInfo() 함수로부터 movieId(movie.id) 받아와야 함 -> O
