+++ src/main_src/displayMovies.js
@@ -7,7 +7,7 @@ export const displayMovies = async () => {
   $movieList.innerHTML = movieInfo
     .map(
       (movie) => `
-      <a href="./detailMovie2.html?id=${movie.id}" class="card">
+      <a href="./detailMovie.html?id=${movie.id}" class="card">
         <img src="https://image.tmdb.org/t/p/w500${movie.poster_path}" alt="${movie.title}">
           <h1>${movie.title}</h1>
           <p>평점: ⭐️${movie.vote_average}</p>
